% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pediatric_regensburg_preprocess.R
\name{construct_dataset_pediatric_regensburg}
\alias{construct_dataset_pediatric_regensburg}
\title{Pediatric_regensburg_preprocess}
\usage{
construct_dataset_pediatric_regensburg()
}
\description{
Prepare the pediatric_regensburg dataset for the DNCIT package
}
